refactor: socket on/emit 타입 파일분리
